Manual diff of 3d-tom branch. Includes rebase from master. Moving of deps and turtle* files into lib.
