+++ .travis.yml
@@ -1,0 +1,5 @@
+language: python
+python:
+  - "2.7"
+install: pip install -r requirements.txt --use-mirrors
+script:  nosetests spec/*
