+++ requirements.txt
@@ -1,0 +1,7 @@
+argparse==1.2.1
+mock==1.0.1
+nose==1.3.1
+noseOfYeti==1.4.9
+python-termstyle==0.1.10
+rednose==0.4.1
+wsgiref==0.1.2
