+++ setup.cfg
@@ -1,0 +1,3 @@
+[nosetests]
+with-noy=1
+rednose=1
